+++ src/index.ts
@@ -1,5 +1,5 @@
 import packageJSON from '../package.json' with {type: 'json'};
-import {Map, type MapOptions, type WebGLContextAttributesWithType} from './ui/map';
+import {Map, type MapOptions, type WebGLContextAttributesWithType, type CoveringTilesOptions} from './ui/map';
 import {NavigationControl, type NavigationControlOptions} from './ui/control/navigation_control';
 import {GeolocateControl, type GeolocateControlOptions} from './ui/control/geolocate_control';
 import {AttributionControl, type AttributionControlOptions} from './ui/control/attribution_control';
@@ -365,5 +365,6 @@ export {
     removeProtocol,
     addSourceType,
     importScriptInWorkers,
-    createTileMesh
+    createTileMesh,
+    CoveringTilesOptions
 };